addressing review comments, refactor: consolidate common packages and fix service ownership logic, Remove outdated test cases and comments
