+++ controllers/refresolver/lbservice/lbservice.go
@@ -23,7 +23,7 @@ import (
 	"fmt"
 
 	k8gbv1beta1 "github.com/k8gb-io/k8gb/api/v1beta1"
-	"github.com/k8gb-io/k8gb/controllers/refresolver/lbservice/common"
+	"github.com/k8gb-io/k8gb/controllers/refresolver/common"
 	"github.com/k8gb-io/k8gb/controllers/utils"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
